test: consolidate mysql and sqlite family cases
